Typescriptify prune
