+++ src/store/cache/users/utils/index.ts
@@ -1,2 +1,3 @@
 export { reformat } from './reformat'
 export { processAndCacheUsers } from './processAndCacheUsers'
+export { pruneBlobValues } from './pruneBlobValues'
